// diet-app-backend/src/schemas/diets.ts
import z from 'zod'

export const DietPlanRequestSchema = z.object({
  name: z.string().min(2),
  weight: z.number().positive(),
  height: z.number().positive(),
  age: z.number().positive().int(),
  activity_level: z.enum(['sedentario', '2x_semana', '4x_semana']),
  genre: z.enum(['masculino', 'feminino', 'outro']),
  goal: z.enum(['perda_de_peso', 'hipertrofia', 'manter_massa_muscular']),
})

export type DietPlanRequest = z.infer<typeof DietPlanRequestSchema>